src/foodie-app/requirements.txt
Flask
requests
pymongo
python-dotenv
dnspython